Refactored the babe pallet
